Added frobbing and frobcubes
Player can now frob gameobjects that are either marked in the source code, or have an OnFrob script call. Also added a frobcube that can be frobbed by the player. Player cannot frob frobcube if it is not solid, needs to be looked into. Also added frobcube gameobject to the engine def files.